-use size_t -recheckin header as unix file
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@103684 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/dba/libflatfile/flatfile.h
@@ -23,14 +23,14 @@
 
 typedef struct {
 	char *dptr;
-	int dsize;
+	size_t dsize;
 } datum;
 
 typedef struct {
 	char *lockfn;
 	int lockfd;
 	php_stream *fp;
-	long CurrentFlatFilePos;
+	size_t CurrentFlatFilePos;
 	datum nextkey;
 } flatfile;
